Move NearInteractionTouchable inspector to separate file and expose some properties for testing.
