+++ Assets/MixedRealityToolkit.Services/InputSystem/Inspectors/NearInteractionTouchableInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee3aaf2ee89fde64685e5b0606668d08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
