+++ Assets/MixedRealityToolkit.Services/InputSystem/NearInteractionTouchable.cs
@@ -15,118 +15,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
     /// </summary>
     public class NearInteractionTouchable : BaseNearInteractionTouchable
     {
-#if UNITY_EDITOR
-        [UnityEditor.CustomEditor(typeof(NearInteractionTouchable))]
-        public class Editor : UnityEditor.Editor
-        {
-            private readonly Color handleColor = Color.white;
-            private readonly Color fillColor = new Color(0, 0, 0, 0);
-
-            protected virtual void OnSceneGUI()
-            {
-                NearInteractionTouchable t = (NearInteractionTouchable)target;
-
-                if (Event.current.type == EventType.Repaint)
-                {
-                    UnityEditor.Handles.color = handleColor;
-
-                    Vector3 center = t.transform.TransformPoint(t.localCenter);
-
-                    float arrowSize = UnityEditor.HandleUtility.GetHandleSize(center) * 0.75f;
-                    UnityEditor.Handles.ArrowHandleCap(0, center, Quaternion.LookRotation(t.transform.rotation * t.localForward), arrowSize, EventType.Repaint);
-
-                    Vector3 rightDelta = t.transform.localToWorldMatrix.MultiplyVector(t.LocalRight * t.bounds.x / 2);
-                    Vector3 upDelta = t.transform.localToWorldMatrix.MultiplyVector(t.localUp * t.bounds.y / 2);
-
-                    Vector3[] points = new Vector3[4];
-                    points[0] = center + rightDelta + upDelta;
-                    points[1] = center - rightDelta + upDelta;
-                    points[2] = center - rightDelta - upDelta;
-                    points[3] = center + rightDelta - upDelta;
-
-                    UnityEditor.Handles.DrawSolidRectangleWithOutline(points, fillColor, handleColor);
-                }
-            }
-
-            public override void OnInspectorGUI()
-            {
-                base.OnInspectorGUI();
-
-                NearInteractionTouchable t = (NearInteractionTouchable)target;
-                BoxCollider bc = t.GetComponent<BoxCollider>();
-                RectTransform rt = t.GetComponent<RectTransform>();
-                if (bc != null)
-                {
-                    // project size to local coordinate system
-                    Vector2 adjustedSize = new Vector2(
-                                Math.Abs(Vector3.Dot(bc.size, t.LocalRight)),
-                                Math.Abs(Vector3.Dot(bc.size, t.localUp)));
-
-                    // Resize helper
-                    if (adjustedSize != t.bounds)
-                    {
-                        UnityEditor.EditorGUILayout.HelpBox("Bounds do not match the BoxCollider size", UnityEditor.MessageType.Warning);
-                        if (GUILayout.Button("Fix Bounds"))
-                        {
-                            UnityEditor.Undo.RecordObject(t, "Fix Bounds");
-                            t.bounds = adjustedSize;
-                        }
-                    }
-
-                    // Recentre helper
-                    if (t.localCenter != bc.center + Vector3.Scale(bc.size / 2.0f, t.localForward))
-                    {
-                        UnityEditor.EditorGUILayout.HelpBox("Center does not match the BoxCollider center", UnityEditor.MessageType.Warning);
-                        if (GUILayout.Button("Fix Center"))
-                        {
-                            UnityEditor.Undo.RecordObject(t, "Fix Center");
-                            t.localCenter = bc.center + Vector3.Scale(bc.size / 2.0f, t.localForward);
-                        }
-                    }
-                }
-                else if (rt != null)
-                {
-                    // Resize Helper
-                    if (rt.sizeDelta != t.bounds)
-                    {
-                        UnityEditor.EditorGUILayout.HelpBox("Bounds do not match the RectTransform size", UnityEditor.MessageType.Warning);
-                        if (GUILayout.Button("Fix Bounds"))
-                        {
-                            UnityEditor.Undo.RecordObject(t, "Fix Bounds");
-                            t.bounds = rt.sizeDelta;
-                        }
-                    }
-
-                    if (t.GetComponentInParent<Canvas>() != null && t.localForward != new Vector3(0, 0, -1))
-                    {
-                        UnityEditor.EditorGUILayout.HelpBox("Unity UI generally has forward facing away from the front. The LocalForward direction specified does not match the expected forward direction.", UnityEditor.MessageType.Warning);
-                        if (GUILayout.Button("Fix Forward Direction"))
-                        {
-                            UnityEditor.Undo.RecordObject(t, "Fix Forward Direction");
-                            t.localForward = new Vector3(0, 0, -1);
-                        }
-                    }
-                }
-
-                // Perpendicular forward/up vectors helpers
-                if (Vector3.Dot(t.localForward, t.localUp) != 0)
-                {
-                    UnityEditor.EditorGUILayout.HelpBox("Local Forward and Local Up are not perpendicular.", UnityEditor.MessageType.Warning);
-                    if (GUILayout.Button("Fix Local Up"))
-                    {
-                        UnityEditor.Undo.RecordObject(t, "Fix Local Up");
-                        t.localUp = Vector3.Cross(t.localForward, t.LocalRight).normalized;
-                    }
-                    if (GUILayout.Button("Fix Local Forward"))
-                    {
-                        UnityEditor.Undo.RecordObject(t, "Fix Local Forward");
-                        t.localForward = Vector3.Cross(t.LocalRight, t.localUp).normalized;
-                    }
-                }
-            }
-        }
-#endif
-
         private enum TouchableSurface
         {
             BoxCollider,
@@ -147,7 +35,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
         /// </summary>
         [SerializeField]
         protected Vector3 localUp = Vector3.up;
-
         public Vector3 LocalUp { get => localUp; }
 
         /// <summary>
@@ -155,6 +42,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
         /// </summary>
         [SerializeField]
         protected Vector3 localCenter = Vector3.zero;
+        public Vector3 LocalCenter { get => localCenter; set { localCenter = value; } }
 
         [SerializeField]
         [Tooltip("The type of surface to calculate the touch point on.")]
@@ -179,11 +67,24 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
         public Vector3 Forward => transform.TransformDirection(localForward);
 
+        public void SetLocalForward(Vector3 newLocalForward)
+        {
+            localForward = newLocalForward;
+            localUp = Vector3.Cross(localForward, LocalRight).normalized;
+        }
+
+        public void SetLocalUp(Vector3 newLocalUp)
+        {
+            localForward = newLocalUp;
+            localUp = Vector3.Cross(localForward, LocalRight).normalized;
+        }
+
         /// <summary>
         /// Local space forward direction
         /// </summary>
         [SerializeField]
         protected Vector2 bounds = Vector2.zero;
+        public Vector2 Bounds { get => bounds; set { bounds = value; } }
 
         protected new void OnValidate()
         {
